requirements.txt
tflearn==0.3.2
tensorflow-gpu==1.14.0
numpy==1.14.6
glob2==0.7
opencv-python==4.1.1.26
Pillow==6.1.0
image-slicer==0.3.0
numpy==1.14.6